布局优化，即一些细节的处理

+++ app/src/main/java/com/example/wuxinmima/mimaxinxixiugai.java
@@ -8,6 +8,7 @@ import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
+import android.widget.Toast;
 
 import org.litepal.LitePal;
 
@@ -30,26 +31,37 @@ public class mimaxinxixiugai extends AppCompatActivity {
         修改.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                mima mima1 = new mima();
-                Intent intent = getIntent();
-                String 用户名 = intent.getStringExtra("用户名");
-                String 名称 = intent.getStringExtra("名称");
-                LitePal.deleteAll(mima.class,"名称=?",名称);
+
                 String 新密码 = 修改密码.getText().toString();
                 String 新账号 = 修改账号.getText().toString();
                 String 新绑定手机号 = 修改绑定手机号.getText().toString();
-                mima1.set用户名(用户名);
-                mima1.set名称(名称);
-                mima1.set账号(新账号);
-                mima1.set密码(新密码);
-                mima1.set绑定手机号(新绑定手机号);
-                mima1.save();
-                Intent intent1 = new Intent();
-                intent1.putExtra("新密码",新密码);
-                intent1.putExtra("新账号",新账号);
-                intent1.putExtra("新绑定手机号",新绑定手机号);
-                setResult(RESULT_OK,intent1);
-                finish();
+                if (新账号.isEmpty()){
+                    新账号 = "无";
+                }
+                if (新绑定手机号.isEmpty()){
+                    新绑定手机号 = "无";
+                }
+                if (新密码.isEmpty()) {
+                    Toast.makeText(mimaxinxixiugai.this,"请输入新密码",Toast.LENGTH_SHORT).show();
+                }else{
+                    mima mima1 = new mima();
+                    Intent intent = getIntent();
+                    String 用户名 = intent.getStringExtra("用户名");
+                    String 名称 = intent.getStringExtra("名称");
+                    LitePal.deleteAll(mima.class,"名称=?",名称);
+                    mima1.set用户名(用户名);
+                    mima1.set名称(名称);
+                    mima1.set账号(新账号);
+                    mima1.set密码(新密码);
+                    mima1.set绑定手机号(新绑定手机号);
+                    mima1.save();
+                    Intent intent1 = new Intent();
+                    intent1.putExtra("新密码",新密码);
+                    intent1.putExtra("新账号",新账号);
+                    intent1.putExtra("新绑定手机号",新绑定手机号);
+                    setResult(RESULT_OK,intent1);
+                    finish();
+                }
             }
         });
         取消.setOnClickListener(new View.OnClickListener() {
